+++ Assets/OxGFrame/AssetLoader/Scripts/Runtime/Bundle/PatchEvent/PatchEvents.cs
@@ -122,7 +122,7 @@ namespace OxGFrame.AssetLoader.PatchEvent
 
             private static long _lastDownloadSizeBytes;
 
-            public static async void SendEventMessage(int totalDownloadCount, int currentDownloadCount, long totalDownloadSizeBytes, long currentDownloadSizeBytes)
+            public static void SendEventMessage(int totalDownloadCount, int currentDownloadCount, long totalDownloadSizeBytes, long currentDownloadSizeBytes)
             {
                 var msg = new PatchDownloadProgression();
                 msg.totalDownloadCount = totalDownloadCount;
@@ -132,7 +132,6 @@ namespace OxGFrame.AssetLoader.PatchEvent
                 msg.downloadSpeedSizeBytes = msg.currentDownloadSizeBytes - _lastDownloadSizeBytes;
                 _lastDownloadSizeBytes = msg.currentDownloadSizeBytes;
                 msg.progress = currentDownloadSizeBytes * 1f / totalDownloadSizeBytes * 1f;
-                await UniTask.Delay(TimeSpan.FromSeconds(1f));
                 UniEvent.SendMessage(msg);
             }
         }